Fix more bugs then before
